Use 64-bit offsets for Power MemoryReference
Previously, the Power implementation of TR::MemoryReference used 32-bit
offsets in most places. However, this is not sufficiently long to
contain the 34-bit signed offsets that can be used with the new prefixed
load and store instructions introduced in POWER10. In order to properly
represent such offsets, the offset is now represented as an int64_t.

Unfortunately, it's not possible to use a simple constructor for the new
larger offsets due to issues with ambiguity between the overloaded
constructors. Trying to replace the existing int32_t version with only
an int64_t version results in ambiguities when passing the constant 0,
which can be converted to either int32_t or int64_t. Attempting to add
two different constructors for int32_t and int64_t results in a lot of
ambiguities when deciding which should be used (should uint32_t use the
old one or the new one?) which would require an overload for every
single integral type that might end up being passed in.

In order to resolve these issues, a new static factory method called
withDisplacement has been added which provides access to the new
constructor. The new constructor then has an extra useless int argument
to make sure it's never ambiguous with the old one.

+++ compiler/p/codegen/OMRInstruction.hpp
@@ -87,7 +87,7 @@ class OMR_EXTENSIBLE Instruction : public OMR::Instruction
 
    virtual TR::Register *getMemoryBase()                     {return NULL;}
    virtual TR::Register *getMemoryIndex()                    {return NULL;}
-   virtual int32_t      getOffset()                         {return 0;}
+   virtual int64_t      getOffset()                         {return 0;}
    virtual bool  usesCountRegister()                        {return _opcode.usesCountRegister();}
    virtual bool  setsCountRegister();
